DIV-3300 - renamed loadMiniPetition to petitionMiddleware
To avoid loadMiniPetition.loadMiniPetition
