+++ steps/respond/Respond.step.js
@@ -2,7 +2,7 @@ const { Interstitial } = require('@hmcts/one-per-page/steps');
 const { goTo } = require('@hmcts/one-per-page/flow');
 const config = require('config');
 const idam = require('services/idam');
-const loadMiniPetition = require('middleware/loadMiniPetition');
+const petitionMiddleware = require('middleware/petitionMiddleware');
 
 class Respond extends Interstitial {
   static get path() {
@@ -26,7 +26,7 @@ class Respond extends Interstitial {
     return [
       ...super.middleware,
       idam.protect(),
-      loadMiniPetition.loadMiniPetition
+      petitionMiddleware.loadMiniPetition
     ];
   }
 }
